+++ shard/shard.go
@@ -1,13 +1,15 @@
-package util
+package shard
 
 import (
 	"crypto/md5"
 	"encoding/binary"
 	"math"
+
+	"github.com/grafana/dskit/yolo"
 )
 
+// Sharding Strategies.
 const (
-	// Sharding strategies.
 	ShardingStrategyDefault = "default"
 	ShardingStrategyShuffle = "shuffle-sharding"
 )
@@ -20,10 +22,10 @@ var (
 func ShuffleShardSeed(identifier, zone string) int64 {
 	// Use the identifier to compute an hash we'll use to seed the random.
 	hasher := md5.New()
-	hasher.Write(YoloBuf(identifier)) // nolint:errcheck
+	hasher.Write(yolo.Buf(identifier)) // nolint:errcheck
 	if zone != "" {
-		hasher.Write(seedSeparator) // nolint:errcheck
-		hasher.Write(YoloBuf(zone)) // nolint:errcheck
+		hasher.Write(seedSeparator)  // nolint:errcheck
+		hasher.Write(yolo.Buf(zone)) // nolint:errcheck
 	}
 	checksum := hasher.Sum(nil)