+++ shard/shard_test.go
@@ -1,4 +1,4 @@
-package util
+package shard
 
 import (
 	"testing"